Use absolute path when parsing a file in a project
We were passing the project-relative path of a file to javac when working with Java files
in Java projects.
The reason it was working before was because we were directly sending
the file's contents to the compiler.

This change fixes an error marker I get when opening any module-info.java,
where javac would complain that the file is not on the source path.
This is because javac was told the file was located at `/src/module-info.java`,
but the source path it got was correctly set to (eg.)
`/home/davthomp/Projects/module-project/src`.

Signed-off-by: David Thompson <davthomp@redhat.com>