Completion Perf: Delay `scrapeAccessibleBindings`
scrapeAccessibleBindings may run some search and queries that can be
time consuming (~10% of completion time), however, the results of those
queries are not always useful.
So instead of forcing the computation of binding early, we allow to mark
them as "requested" so that they would be computed lazily, typically
when applying a method completion proposal.